Convert to controlled component using hooks
I imported the useState hook and created a formState object comprised of
all the input fields. After that I added value attributes to each input
field and tied it to its matching formState value. This was done to make
the form a controlled component. I also added an onChange event to each
input field so that the displayed value will update as the user types.
Lastly, I added an onSubmit event to the form element which invokes
event.preventDefault.